+++ tests/conftest.py
@@ -22,7 +22,7 @@ def set_env(monkeypatch):
     monkeypatch.setenv("AWS_REGION", "us-west-2")
     monkeypatch.delenv("AWS_PROFILE", raising=False)
     monkeypatch.setenv("AWS_CONFIG_FILE", "/tmp/noconfigheere")
-    monkeypatch.setenv("API_CACHECONTROL", "private, max-age=3600")
+    monkeypatch.setenv("TITILER_API_CACHECONTROL", "private, max-age=3600")
 
 
 @pytest.fixture(autouse=True)