+++ CHANGES.md
@@ -13,6 +13,7 @@
 * add `rio_tiler.io.MultiBandReader` bands dependencies (https://github.com/developmentseed/titiler/pull/226)
 * add `MultiBaseTilerFactory` and `MultiBandTilerFactory` custom tiler factories (https://github.com/developmentseed/titiler/pull/230)
 * Update STAC tiler to use the new `MultiBaseTilerFactory` factory
+* depreciate *empty* GET endpoint for MosaicTilerFactory read (https://github.com/developmentseed/titiler/pull/232)
 
 **breaking change**