+++ CHANGES.md
@@ -19,6 +19,7 @@
 
 * update `titiler.dependencies.AssetsBidxParams` to make `asset` a required parameter (https://github.com/developmentseed/titiler/pull/230
 * the STAC `/info` endpoint now expect the `assets` parameter to be passed. To ge the list of available assets we added a new `/assets` endpoint within the tiler factory
+* remove `COGReader` as default `reader` in `titiler.endpoints.factory.BaseTilerFactory`
 
 ## 0.1.0a14 (2021-01-05)